+++ serde-encrypt/tests/feat_serde_types.rs
@@ -15,11 +15,9 @@ use serde_encrypt::{
     serialize::{impls::CborSerializer, TypedSerialized},
     shared_key::SharedKey,
     traits::{SerdeEncryptPublicKey, SerdeEncryptSharedKey},
+    AsSharedKey,
 };
-use serde_encrypt_core::{
-    error::{Error, ErrorKind},
-    key::as_shared_key::AsSharedKey,
-};
+use serde_encrypt_core::error::{Error, ErrorKind};
 use test_util::{serde_encrypt_public_key::*, serde_encrypt_shared_key::*};
 use void::Void;