+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
   - `serde_encrypt::ReceiverCombinedKey`
   - `serde_encrypt::SenderKeyPair`
   - `serde_encrypt::ReceiverKeyPair`
+  - `serde_encrypt::AsSharedKey`
 
 ---